+++ src/main/java/com/hyunn/capstone/dto/request/ThreeDimensionCreateRequest.java
@@ -1,4 +1,4 @@
-package com.hyunn.capstone.dto.Request;
+package com.hyunn.capstone.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
