+++ src/main/java/com/hyunn/capstone/dto/response/MessageResponse.java
@@ -1,4 +1,4 @@
-package com.hyunn.capstone.dto.Response;
+package com.hyunn.capstone.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
