+++ src/main/java/com/hyunn/capstone/exception/handler/MessageExceptionController.java
@@ -1,4 +1,4 @@
-package com.hyunn.capstone.exception.Handler;
+package com.hyunn.capstone.exception.handler;
 
 import static com.hyunn.capstone.exception.ErrorStatus.INVALID_JSON_EXCEPTION;
 import static com.hyunn.capstone.exception.ErrorStatus.INVALID_PARAMETER;
@@ -6,10 +6,10 @@ import static com.hyunn.capstone.exception.ErrorStatus.MEDIA_TYPE_NOT_SUPPORTED_
 import static com.hyunn.capstone.exception.ErrorStatus.NEED_MORE_PARAMETER;
 import static com.hyunn.capstone.exception.ErrorStatus.VALIDATION_EXCEPTION;
 
-import com.hyunn.capstone.controller.UserController;
-import com.hyunn.capstone.dto.Response.ApiStandardResponse;
-import com.hyunn.capstone.dto.Response.ErrorResponse;
-import com.hyunn.capstone.exception.ImageNotFoundException;
+import com.hyunn.capstone.controller.MessageController;
+import com.hyunn.capstone.dto.response.ApiStandardResponse;
+import com.hyunn.capstone.dto.response.ErrorResponse;
+import com.hyunn.capstone.exception.ApiNotFoundException;
 import com.hyunn.capstone.exception.RootUserException;
 import com.hyunn.capstone.exception.UserNotFoundException;
 import jakarta.validation.ConstraintViolation;
@@ -30,8 +30,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
 @Slf4j
-@RestControllerAdvice(assignableTypes = {UserController.class})
-public class UserExceptionHandler {
+@RestControllerAdvice(assignableTypes = {MessageController.class})
+public class MessageExceptionController {
 
   // 루트 계정에 허가되지 않은 조작을 하는 경우
   @ExceptionHandler(RootUserException.class)
@@ -43,20 +43,20 @@ public class UserExceptionHandler {
     return ApiStandardResponse.fail(errorResponse);
   }
 
-  // 유저를 찾을 수 없는 경우
-  @ExceptionHandler(UserNotFoundException.class)
+  // API 응답이 올바르지 않은 경우
+  @ExceptionHandler(ApiNotFoundException.class)
   @ResponseStatus(HttpStatus.NOT_FOUND)
-  public ApiStandardResponse<ErrorResponse> handleUserNotFoundException(UserNotFoundException e) {
+  public ApiStandardResponse<ErrorResponse> handleApiNotFoundException(ApiNotFoundException e) {
     log.error("", e);
 
     final ErrorResponse errorResponse = ErrorResponse.create(e.toErrorCode(), e.getMessage());
     return ApiStandardResponse.fail(errorResponse);
   }
 
-  // 이미지를 찾을 수 없는 경우
-  @ExceptionHandler(ImageNotFoundException.class)
+  // 유저를 찾을 수 없는 경우
+  @ExceptionHandler(UserNotFoundException.class)
   @ResponseStatus(HttpStatus.NOT_FOUND)
-  public ApiStandardResponse<ErrorResponse> handleImageNotFoundException(ImageNotFoundException e) {
+  public ApiStandardResponse<ErrorResponse> handleUserNotFoundException(UserNotFoundException e) {
     log.error("", e);
 
     final ErrorResponse errorResponse = ErrorResponse.create(e.toErrorCode(), e.getMessage());
