+++ src/main/java/com/hyunn/capstone/service/ImageService.java
@@ -6,7 +6,7 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hyunn.capstone.dto.Response.ImageToTextResponse;
+import com.hyunn.capstone.dto.response.ImageToTextResponse;
 import com.hyunn.capstone.exception.ApiKeyNotValidException;
 import com.hyunn.capstone.exception.ApiNotFoundException;
 import com.hyunn.capstone.exception.FileNotAllowedException;
