+++ src/main/java/com/hyunn/capstone/service/MessageService.java
@@ -3,8 +3,8 @@ package com.hyunn.capstone.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hyunn.capstone.dto.Request.MessageRequest;
-import com.hyunn.capstone.dto.Response.MessageResponse;
+import com.hyunn.capstone.dto.request.MessageRequest;
+import com.hyunn.capstone.dto.response.MessageResponse;
 import com.hyunn.capstone.entity.User;
 import com.hyunn.capstone.exception.ApiKeyNotValidException;
 import com.hyunn.capstone.exception.ApiNotFoundException;
